refactor(frontend-demo): decompose home page into components
Split app/page.tsx into app/components/home/* sections and move
GITHUB_REPO and SYNTAX to app/lib/constants.ts. No visual or logic changes.

Made-with: Cursor
