+++ frontend-demo/app/components/home/WhatIs.tsx
@@ -1,0 +1,94 @@
+function PipelineStep({
+  num,
+  title,
+  desc,
+}: {
+  num: string;
+  title: string;
+  desc: string;
+}) {
+  return (
+    <div className="flex items-start gap-3.5 border-b border-border py-3.5">
+      {/* Step badge: sub-20% alpha on accent — kept inline for exact rgba match */}
+      <div
+        className="flex size-7 shrink-0 items-center justify-center rounded-md border font-mono text-xs font-bold text-accent"
+        style={{
+          background: "rgba(0,229,160,0.1)",
+          borderColor: "rgba(0,229,160,0.2)",
+        }}
+      >
+        {num}
+      </div>
+      <div>
+        <h4 className="mb-0.5 text-[0.9rem] font-medium text-foreground">
+          {title}
+        </h4>
+        <p className="m-0 text-[0.82rem] text-muted">{desc}</p>
+      </div>
+    </div>
+  );
+}
+
+export default function WhatIs() {
+  const steps = [
+    {
+      num: "01",
+      title: "Ingest",
+      desc: "Load PDFs, HTML, text files. Auto-chunked and embedded.",
+    },
+    {
+      num: "02",
+      title: "Index",
+      desc: "FAISS, Chroma, or your custom vector store. Your choice.",
+    },
+    {
+      num: "03",
+      title: "Retrieve",
+      desc: "Semantic search with MMR re-ranking for diversity.",
+    },
+    {
+      num: "04",
+      title: "Generate",
+      desc: "LLM answer grounded in retrieved context. Cited.",
+    },
+  ];
+  return (
+    <section id="about" className="bg-background px-8 py-24">
+      <div className="mx-auto max-w-[1100px]">
+        <p className="mb-4 font-mono text-[0.78rem] uppercase tracking-[2px] text-accent">
+          {"// what is chatvector"}
+        </p>
+        <h2 className="mb-5 text-[clamp(1.8rem,3.5vw,2.8rem)] font-semibold leading-tight tracking-[-0.8px] text-foreground">
+          RAG that&apos;s sharp, fast,
+          <br />
+          and open source.
+        </h2>
+        <p className="max-w-[560px] text-[1.05rem] font-light leading-[1.7] text-muted">
+          ChatVector handles the entire retrieval pipeline — from raw documents
+          to grounded LLM responses — so you can focus on building, not
+          plumbing.
+        </p>
+
+        <div className="mt-12 grid grid-cols-1 items-center gap-12 md:grid-cols-2">
+          <div>
+            <p className="mb-5 text-[0.95rem] leading-[1.8] text-muted">
+              Most RAG implementations are fragile, slow, or locked into a
+              vendor. ChatVector is different — a clean, composable engine built
+              for developers who want full control.
+            </p>
+            <p className="text-[0.95rem] leading-[1.8] text-muted">
+              Swap your vector store, your LLM, or your chunking strategy
+              without rewriting your app. Built on battle-tested primitives.
+              Runs anywhere Python runs.
+            </p>
+          </div>
+          <div className="rounded-xl border border-border bg-surface p-6">
+            {steps.map((s) => (
+              <PipelineStep key={s.num} {...s} />
+            ))}
+          </div>
+        </div>
+      </div>
+    </section>
+  );
+}
