+++ frontend-demo/app/lib/constants.ts
@@ -1,0 +1,10 @@
+export const GITHUB_REPO = "https://github.com/chatvector-ai/chatvector-ai";
+
+/** Code-sample syntax colors (IDE-style); not part of the seven semantic tokens. */
+export const SYNTAX = {
+  kw: "rgb(255, 123, 114)",
+  fn: "rgb(121, 192, 255)",
+  str: "rgb(165, 214, 255)",
+  cm: "rgb(139, 148, 158)",
+  plain: "rgb(201, 209, 217)",
+} as const;
